+++ server/database.go
@@ -448,6 +448,10 @@ func (db *Database) migrate() error {
 		return formatError(err, "")
 	}
 
+	if err := migrateUserAlertPushPreferences(db); err != nil {
+		return formatError(err, "")
+	}
+
 	return nil
 }